+++ packages/ui/react-ui-form/src/components/Form/Form.theme.ts
@@ -19,9 +19,11 @@ const formStyles = tv({
     sectionDescription: 'text-description',
     fieldSet: '',
     field: '',
-    description: 'text-description',
-    control: '',
-    validation: '',
+    fieldLabel: 'grid grid-cols-[1fr_auto_auto] items-center select-none',
+    fieldLabelText: '',
+    fieldDescription: 'text-description',
+    fieldControl: '',
+    fieldValidation: '',
   },
   variants: {
     variant: {
@@ -34,14 +36,15 @@ const formStyles = tv({
         fieldSet: 'flex flex-col gap-trim-md pt-trim-md',
         field: mx(
           'grid',
-          'grid-cols-1 [grid-template-areas:"header""description""control"]',
-          'md:grid-cols-2 md:[grid-template-areas:"header_control""description_description"]',
+          'grid-cols-1 [grid-template-areas:"header""description""control""validation"]',
+          'md:grid-cols-2 md:[grid-template-areas:"header_control""description_description""validation_validation"]',
           'gap-x-trim-lg gap-y-0 p-trim-md border border-separator rounded-sm',
         ),
-        description: '[grid-area:description] text-base text-description',
-        control: '[grid-area:control] flex items-center md:justify-end pt-3 md:pt-0',
-        // The error block is outside the named areas; span it full-width so `layout='full'` settings forms don't misplace it.
-        validation: 'col-span-full',
+        fieldLabel: '[grid-area:header]',
+        fieldLabelText: 'pb-trim-md text-base-fg text-lg',
+        fieldDescription: '[grid-area:description] text-base text-description',
+        fieldControl: '[grid-area:control] flex items-center md:justify-end pt-3 md:pt-0',
+        fieldValidation: '[grid-area:validation]',
       },
     },
   },
