+++ packages/ui/react-ui-form/src/components/Form/FormField/FormFieldWrapper.tsx
@@ -6,14 +6,90 @@ import { format as formatDate } from 'date-fns';
 import React, { Component, type PropsWithChildren, type ReactNode } from 'react';
 
 import { Format } from '@dxos/echo';
-import { Input } from '@dxos/react-ui';
+import { inputTextLabel, Icon, Input, type ThemedClassName, Tooltip } from '@dxos/react-ui';
+import { mx } from '@dxos/ui-theme';
 
 import { type FormFieldRendererProps } from '#types';
 
 import { useFormContext } from '../../../hooks';
-import { formTheme } from '../Form.theme';
-import { FormFieldLabel } from './FormFieldLabel';
+import { type FormVariant, formTheme } from '../Form.theme';
 import { type FieldPresentation, presentationFor } from './presentation';
+
+//
+// FormFieldLabel
+//
+
+export type FormFieldLabelProps = ThemedClassName<
+  {
+    /** Render a plain `<span>` instead of an input-associated `Input.Label`, for labels used outside an `Input.Root` (e.g. section/group headers). */
+    standalone?: boolean;
+    /** Class applied to the inner label text node, overriding the default size/color (e.g. `text-lg`). */
+    labelClassName?: string;
+    /** Form variant; selects the label's chrome (e.g. `settings` enlarges the text and places it in the field grid). */
+    variant?: FormVariant;
+    error?: string;
+    /**
+     * JSON path of the field this label describes (e.g. `runtime.client.storage.persistent`).
+     * Field metadata; accepted by callers but not currently rendered.
+     */
+    path?: string;
+    /**
+     * Trailing button rendered at the end of the label row (third grid column).
+     * Used by nested-object field sets to surface a collapse toggle.
+     */
+    button?: ReactNode;
+    onClick?: () => void;
+  } & Pick<FormFieldRendererProps, 'label' | 'readonly' | 'required'>
+>;
+
+export const FormFieldLabel = ({
+  classNames,
+  labelClassName,
+  variant = 'default',
+  label,
+  error,
+  readonly,
+  required,
+  standalone,
+  button,
+  onClick,
+}: FormFieldLabelProps) => {
+  const styles = formTheme.styles({ variant });
+  // Render the required asterisk via a `::after` pseudo-element rather than a DOM node: it keeps the
+  // label's `textContent` exactly `label`, so fields stay locatable by their exact label text
+  // (`getByLabelText('Name')`), which the DOM-text-based query would otherwise miss as `Name *`.
+  // The `fieldLabelText` slot is applied last so a variant/caller size/color (e.g. `text-lg`) wins over
+  // the `text-sm`/`text-description` baked into `inputTextLabel`.
+  const labelClassNames = mx(
+    inputTextLabel,
+    required && "after:content-['*'] after:ms-0.5 after:text-warning-text",
+    styles.fieldLabelText({ class: labelClassName }),
+  );
+  // `Input.Label` is a themed primitive that reads `classNames` (and ignores `className`), whereas the
+  // plain `span` used for read-only/standalone labels reads `className`.
+  const labelNode =
+    readonly || standalone ? (
+      <span className={labelClassNames}>{label}</span>
+    ) : (
+      <Input.Label classNames={labelClassNames}>{label}</Input.Label>
+    );
+
+  return (
+    <div className={styles.fieldLabel({ class: mx(onClick && 'cursor-pointer', classNames) })} onClick={onClick}>
+      {labelNode}
+      {error ? (
+        <Tooltip.Trigger asChild content={error} side='bottom'>
+          <Icon icon='ph--warning--regular' size={4} classNames='text-error-text' />
+        </Tooltip.Trigger>
+      ) : (
+        <span />
+      )}
+      {button}
+    </div>
+  );
+};
+
+FormFieldLabel.displayName = 'Form.FieldLabel';
 
 //
 // FormFieldWrapper
@@ -110,10 +186,10 @@ export const FormFieldWrapper = <T,>(props: FormFieldWrapperProps<T>) => {
         )}
         {/* Description */}
         {showDescription && description && (
-          <Input.Description classNames={styles.description()}>{description}</Input.Description>
+          <Input.Description classNames={styles.fieldDescription()}>{description}</Input.Description>
         )}
         {/* Control */}
-        <div className={styles.control()}>
+        <div className={styles.fieldControl()}>
           {resolved.isStatic
             ? (renderStatic?.(value) ?? <p className='truncate min-w-0'>{formatStaticValue(value, format)}</p>)
             : children
@@ -122,7 +198,7 @@ export const FormFieldWrapper = <T,>(props: FormFieldWrapperProps<T>) => {
         </div>
         {/* Error */}
         {resolved.showError && error && (
-          <div className={styles.validation()}>
+          <div className={styles.fieldValidation()}>
             <Input.DescriptionAndValidation>
               <Input.Validation>{error}</Input.Validation>
             </Input.DescriptionAndValidation>
