+++ packages/ui/react-ui-form/src/components/Form/FormField/fields/RefField/InlineRefField.tsx
@@ -16,7 +16,7 @@ import { translationKey } from '#translations';
 
 import { omitId } from '../../../../../util';
 import { FormContent, FormFieldSetController, FormRoot } from '../../../FormControls';
-import { FormFieldLabel } from '../../FormFieldLabel';
+import { FormFieldLabel } from '../../FormFieldWrapper';
 import { presentationFor } from '../../presentation';
 import { type RefFieldProps } from './RefField';
